Add a simple test
* Make the code more testable
* Decouple gRPC related code from the rest
* Use interface for redis client
* Add simple test that runs the publish function, with redis mocked. For
now it only ensures that the function does not return an error or
panics, but ideally it should check what requests are made to redis as well
